+++ src/components/Warning.js
@@ -1,12 +1,12 @@
 import React from "react";
 import styled from "styled-components";
 
-import { warning } from "../common.json";
+import { colors } from "../common.json";
 
 const Container = styled.div`
-  background-color: ${warning.background};
-  color: ${warning.font};
-  border: 1px solid ${warning.border};
+  background-color: ${colors.warning.background};
+  color: ${colors.warning.font};
+  border: 1px solid ${colors.warning.border};
   margin-bottom: 20px;
   padding: 15px;
 `;