+++ src/index.js
@@ -1,11 +1,13 @@
 import Icon from "./components/Icon";
 import Info from "./components/Info";
 import Success from "./components/Success";
+import Tag from "./components/Tag";
 import Warning from "./components/Warning";
 
 export {
   Icon,
   Info,
   Success,
+  Tag,
   Warning
 };