+++ test/Icon.test.js
@@ -4,7 +4,7 @@ import { mount } from "enzyme";
 import { AsyncPath } from "../src/components/Icon";
 
 test("renders without crashing", async () => {
-  const component = mount(<AsyncPath icon="alert" />);
+  const component = mount(<AsyncPath icon="checkmark" />);
   expect(component.find("path")).toHaveLength(0);
 
   await component.instance().componentDidMount();