+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `eslint` and the corresponding `yarn lint` command.
+- `prettier` and the corresponding `yarn prettier` command.
+
+### Changed
+- Pass on any extra props to the root in each display component (allows things like passing `className` as a prop).
 
 ## [0.0.4] - 2018-06-18
 ### Added